chore: update docs + provider
