+++ docs/resources/installer.md
@@ -1,14 +1,14 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "nuon_app_installer Resource - terraform-provider-nuon"
+page_title: "nuon_installer Resource - terraform-provider-nuon"
 subcategory: ""
 description: |-
-  A public installer page for a nuon app.
+  A Nuon installer for one or more apps.
 ---
 
-# nuon_app_installer (Resource)
+# nuon_installer (Resource)
 
-A public installer page for a nuon app.
+A Nuon installer for one or more apps.
 
 
 
@@ -17,16 +17,18 @@ A public installer page for a nuon app.
 
 ### Required
 
-- `app_id` (String) The application ID.
+- `app_ids` (Set of String) App IDs to connect to this installer.
 - `community_url` (String) Community url to a slack or discord, etc.
+- `copyright_markdown` (String) Markdown that rendered in the copyright section.
 - `description` (String) Short description of app.
 - `documentation_url` (String) Documentation url
+- `favicon_url` (String) Favicon url
+- `footer_markdown` (String) Markdown that will be rendered in the footer section.
 - `github_url` (String) GitHub url, link to application.
 - `homepage_url` (String) Homepage url
 - `logo_url` (String) Logo image to display on page.
 - `name` (String) App name to render on install page.
 - `post_install_markdown` (String) Markdown that will be shown to users after a successful install. Supports interpolation.
-- `slug` (String) URL slug to access app from.
 
 ### Optional
