+++ internal/provider/provider.go
@@ -107,7 +107,7 @@ func (p *Provider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewAppResource,
 		NewAppInputResource,
-		NewAppInstallerResource,
+		NewInstallerResource,
 		NewAppSandboxResource,
 		NewAppRunnerResource,
 		NewInstallResource,
